Fix formatting for coordinator profiling aggregation

+++ include/distributed/coordinator.hpp
@@ -411,11 +411,11 @@ public:
       start_time -= offset;
       end_time -= offset;
 
-      size_t start_ms =
+      double start_ms =
           Time::duration_cast<Time::microseconds>(start_time.time_since_epoch()).count() / 1000.0;
-      size_t end_ms =
+      double end_ms =
           Time::duration_cast<Time::microseconds>(end_time.time_since_epoch()).count() / 1000.0;
-      size_t duration_ms =
+      double duration_ms =
           Time::duration_cast<Time::microseconds>(end_time - start_time).count() / 1000.0;
 
       if (start_ms < 0 || end_ms < 0 || duration_ms < 0) {
